consolidate webpack configs
