+++ src/standalone-js/utils/mapbox-gl.js
